t/323-net-frame-extents: wait for window to close before proceeding
This test was flaky because it issued a 'kill' command through i3
and assumed that focus would immediately be different afterwards.

However, the 'kill' command (via x_window_kill) only results in an
xcb_destroy_window() request to X11, not in a focus change.
Only when i3 processes the X11 server response to this request
will focus change back to the window that should be changed with
the 'border normal 0' command.

related to https://github.com/i3/i3/issues/3009